backend/requirements.txt
apify_client==1.9.2
Flask==3.0.3
Flask-Cors==5.0.0
flasgger==0.9.7.1
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
supabase==2.13.0
